requirements.txt
# Core Django and Web Framework
Django>=5.1.7,<5.2
djangorestframework>=3.14.0,<4.0
djangorestframework-simplejwt>=5.3.0,<6.0
django-cors-headers>=4.3.1,<5.0

# ASGI Server
daphne>=4.0.0,<5.0
channels>=4.0.0,<5.0
channels-redis>=4.1.0,<5.0

# Database
psycopg2-binary>=2.9.7,<3.0
dj-database-url>=2.1.0,<3.0

# Redis and Celery
redis>=5.0.1,<6.0
celery>=5.3.4,<6.0
django-redis>=5.4.0,<6.0
django-celery-beat>=2.5.0,<3.0

# Image Processing
Pillow>=10.1.0,<11.0

# Audio Processing
librosa>=0.10.1,<1.0
ffmpeg-python>=0.2.0,<1.0

# HTTP and API
requests>=2.31.0,<3.0
pyfcm>=1.5.4,<2.0

# Utilities
python-dotenv>=1.0.0,<2.0
pytz>=2023.3
sqlparse>=0.4.4,<1.0
xxhash>=3.4.1,<4.0

# Development and Testing (optional for production)
faker>=20.1.0,<21.0
matplotlib>=3.7.2,<4.0

# Testing Libraries
pytest>=7.4.0,<8.0
pytest-django>=4.5.2,<5.0
pytest-cov>=4.1.0,<5.0
pytest-mock>=3.11.1,<4.0
pytest-xdist>=3.3.1,<4.0
factory-boy>=3.3.0,<4.0
freezegun>=1.2.2,<2.0
responses>=0.23.3,<1.0
django-test-plus>=2.2.0,<3.0

# WSGI Server (alternative to Daphne)
gunicorn>=21.2.0,<22.0

# Storage (optional)
django-storages[boto3]>=1.14.2,<2.0